Refactor fieldarray and dictfield
