+++ src/tapis-ui/components/jobs/JobLauncher/DictField.module.scss
@@ -1,0 +1,12 @@
+.field {
+  border: 1px dotted gray;
+  padding: 0.5em 0.5em 0.5em 0.5em;
+  margin-bottom: 1em;
+}
+
+// Overrides some nasty absolute positioning from _forms.scss
+.form-input-override {
+  position: inherit !important;
+  margin-top: inherit !important;
+  margin-left: inherit !important;
+}
